하루 이상의 일정 처리
